requirements.txt
aiokafka==0.7.2
anyio==3.6.1
asgiref==3.4.1
async-generator==1.10
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.4
contextlib2==21.6.0
contextvars==2.4
dataclasses==0.6
fastapi==0.78.0
h11==0.12.0
hazelcast-python-client==5.1
httpcore==0.14.7
httpx==0.22.0
idna==3.3
immutables==0.18
importlib-metadata==4.8.3
kafka-python==2.0.2
pydantic==1.9.1
python-consul==1.1.0
python-dotenv==0.20.0
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.16.0
zipp==3.6.0
